fix failing entrypoint script & mv insert mock records to separate sql

+++ tests/unit/buildspec-scripts/test-helper/src/mock_sql/trigger_defaults.sql
@@ -28,7 +28,7 @@ BEGIN
             SELECT account_name 
             FROM account_dim
             LIMIT random_between(0, (
-                SELECT COUNT(*)
+                SELECT COUNT(*)::INT
                 FROM account_dim
                 WHERE account_name != NEW.account_name
             ))
